Post-Quantum_IoT_Signature_Prototype/primary_scripts/requirements.txt
liboqs-python==0.10.0
pip==24.3.1
psutil==6.1.0
